+++ tests/management/test_manage.py
@@ -5,6 +5,7 @@ import pytest
 from cirrus.management.deployment import (
     Deployment,
 )
+from click.testing import Result
 
 from tests.management.conftest import mock_parameters
 
@@ -72,14 +73,14 @@ def test_get_payload(
     deployment,
     create_records,
 ):
-    for payload_id in create_records["processing"]:
+    for payload_id in create_records["completed"]:
         result = deployment(f"get-payload {payload_id}")
         assert result.exit_code == 0
         assert json.loads(result.stdout.strip())["payload_id"] == payload_id
 
 
 def test_get_state(deployment, create_records):
-    for payload_id in create_records["processing"]:
+    for payload_id in create_records["completed"]:
         result = deployment(f"get-state {payload_id}")
         assert result.exit_code == 0
         output = json.loads(result.stdout.strip())
@@ -142,3 +143,27 @@ def test_manage_show_deployment(deployment, put_parameters):
 def test_call_cli_return_values(deployment, command, expect_exit_zero, put_parameters):
     result = deployment(f"call {command}")
     assert result.exit_code == 0 if expect_exit_zero else result.exit_code != 0
+
+
+def assert_get_records(
+    result: Result,
+    create_records: dict[str, list[str]],
+    state: str,
+):
+    assert result.exit_code == 0
+    output = result.stdout.strip().split("\n")
+    assert len(create_records[state]) == len(output)
+    for payload in output:
+        assert json.loads(payload)["payload_id"] in create_records[state]
+
+
+def test_get_records(deployment, create_records, statedb):
+    result = deployment(
+        "get-records --collection-workflow 'sar-test-panda_test' --state 'COMPLETED'",
+    )
+    assert_get_records(result, create_records, "completed")
+
+    result = deployment(
+        "get-records --collection-workflow 'sar-test-panda_test' --state 'FAILED'",
+    )
+    assert_get_records(result, create_records, "failed")